@@ -31,7 +31,7 @@ public class JniConfigGenerator {
             },
             footerHeading = "%nExample%n",
             footer = {
-                    "  java -cp \"myapp.jar;picocli-4.6.3-SNAPSHOT.jar;picocli-codegen-4.6.3-SNAPSHOT.jar\" " +
+                    "  java -cp \"myapp.jar;picocli-4.7.6-SNAPSHOT.jar;picocli-codegen-4.7.6-SNAPSHOT.jar\" " +
                             "picocli.codegen.aot.graalvm.JniConfigGenerator my.pkg.MyClass"
             },
             mixinStandardHelpOptions = true, sortOptions = false,
