@@ -116,7 +116,7 @@ public class AutoComplete {
                     "",
                     "Example",
                     "-------",
-                    "  java -cp \"myapp.jar;picocli-4.6.3-SNAPSHOT.jar\" \\",
+                    "  java -cp \"myapp.jar;picocli-4.7.6-SNAPSHOT.jar\" \\",
                     "              picocli.AutoComplete my.pkg.MyClass"
             },
             exitCodeListHeading = "%nExit Codes:%n",
@@ -293,12 +293,14 @@ public class AutoComplete {
     }
     private static class CommandDescriptor {
         final String functionName;
+        final String parentFunctionName;
         final String parentWithoutTopLevelCommand;
         final String commandName;
         final CommandLine commandLine;
 
-        CommandDescriptor(String functionName, String parentWithoutTopLevelCommand, String commandName, CommandLine commandLine) {
+        CommandDescriptor(String functionName, String parentFunctionName, String parentWithoutTopLevelCommand, String commandName, CommandLine commandLine) {
             this.functionName = functionName;
+            this.parentFunctionName = parentFunctionName;
             this.parentWithoutTopLevelCommand = parentWithoutTopLevelCommand;
             this.commandName = commandName;
             this.commandLine = commandLine;
@@ -312,7 +314,7 @@ public class AutoComplete {
             "# =======================\n" +
             "#\n" +
             "# Bash completion support for the `%1$s` command,\n" +
-            "# generated by [picocli](http://picocli.info/) version %2$s.\n" +
+            "# generated by [picocli](https://picocli.info/) version %2$s.\n" +
             "#\n" +
             "# Installation\n" +
             "# ------------\n" +
@@ -359,7 +361,10 @@ public class AutoComplete {
             "  alias compopt=complete\n" +
             "\n" +
             "  # Enable bash completion in zsh (see [7])\n" +
-            "  autoload -U +X compinit && compinit\n" +
+            "  # Only initialize completions module once to avoid unregistering existing completions.\n" +
+            "  if ! type compdef > /dev/null; then\n" +
+            "    autoload -U +X compinit && compinit\n" +
+            "  fi\n" +
             "  autoload -U +X bashcompinit && bashcompinit\n" +
             "fi\n" +
             "\n" +
@@ -417,6 +422,41 @@ public class AutoComplete {
             "  done\n" +
             "  echo \"$result\"\n" +
             "}\n" +
+            "\n" +
+            "# compReplyArray generates a list of completion suggestions based on an array, ensuring all values are properly escaped.\n" +
+            "#\n" +
+            "# compReplyArray takes a single parameter: the array of options to be displayed\n" +
+            "#\n" +
+            "# The output is echoed to std_out, one option per line.\n"
+            + "#\n"
+            + "# Example usage:\n"
+            + "# local options=(\"foo\", \"bar\", \"baz\")\n"
+            + "# local IFS=$'\\n'\n"
+            + "# COMPREPLY=($(compReplyArray \"${options[@]}\"))\n" +
+            "function compReplyArray() {\n" +
+            "  declare -a options\n" +
+            "  options=(\"$@\")\n" +
+            "  local curr_word=${COMP_WORDS[COMP_CWORD]}\n" +
+            "  local i\n" +
+            "  local quoted\n" +
+            "  local optionList=()\n" +
+            "\n" +
+            "  for (( i=0; i<${#options[@]}; i++ )); do\n" +
+            "    # Double escape, since we want escaped values, but compgen -W expands the argument\n" +
+            "    printf -v quoted %%q \"${options[i]}\"\n" +
+            "    quoted=\\'${quoted//\\'/\\'\\\\\\'\\'}\\'\n" +
+            "\n" +
+            "    optionList[i]=$quoted\n" +
+            "  done\n" +
+            "\n" +
+            "  # We also have to add another round of escaping to $curr_word.\n" +
+            "  curr_word=${curr_word//\\\\/\\\\\\\\}\n" +
+            "  curr_word=${curr_word//\\'/\\\\\\'}\n" +
+            "\n" +
+            "  # Actually generate completions.\n" +
+            "  local IFS=$'\\n'\n" +
+            "  echo -e \"$(compgen -W \"${optionList[*]}\" -- \"$curr_word\")\"\n" +
+            "}\n" +
             "\n";
 
     private static final String SCRIPT_FOOTER = "" +
@@ -462,7 +502,7 @@ public class AutoComplete {
                         "\n" +
                         "LIBS=path/to/libs\n" +
                         "CP=\"${LIBS}/myApp.jar\"\n" +
-                        "java -cp \"${CP}\" '" + commandLine.getCommand().getClass().getName() + "' $@");
+                        "java -cp \"${CP}\" '" + ((Object) commandLine.getCommand()).getClass().getName() + "' $@");
             }
         } finally {
             if (completionWriter != null) { completionWriter.close(); }
@@ -496,7 +536,7 @@ public class AutoComplete {
 
     private static List<CommandDescriptor> createHierarchy(String scriptName, CommandLine commandLine) {
         List<CommandDescriptor> result = new ArrayList<CommandDescriptor>();
-        result.add(new CommandDescriptor("_picocli_" + scriptName, "", scriptName, commandLine));
+        result.add(new CommandDescriptor("_picocli_" + scriptName, "", "", scriptName, commandLine));
         createSubHierarchy(scriptName, "", commandLine, result);
         return result;
     }
@@ -509,9 +549,12 @@ public class AutoComplete {
             String commandName = entry.getKey(); // may be an alias
             String functionNameWithoutPrefix = bashify(concat("_", parentWithoutTopLevelCommand.replace(' ', '_'), commandName));
             String functionName = concat("_", "_picocli", scriptName, functionNameWithoutPrefix);
+            String parentFunctionName = parentWithoutTopLevelCommand.length() == 0
+                ? concat("_", "_picocli", scriptName)
+                : concat("_", "_picocli", scriptName, bashify(parentWithoutTopLevelCommand.replace(' ', '_')));
 
             // remember the function name and associated subcommand so we can easily generate a function later
-            out.add(new CommandDescriptor(functionName, parentWithoutTopLevelCommand, commandName, entry.getValue()));
+            out.add(new CommandDescriptor(functionName, parentFunctionName, parentWithoutTopLevelCommand, commandName, entry.getValue()));
         }
 
         // then recursively do the same for all nested subcommands
@@ -532,6 +575,13 @@ public class AutoComplete {
                 "# on the command line and delegates to the appropriate function\n" +
                 "# to generate possible options and subcommands for the last specified subcommand.\n" +
                 "function _complete_%1$s() {\n" +
+//        # Edge case: if command line has no space after subcommand, then don't assume this subcommand is selected
+//        if [ "${COMP_LINE}" = "${COMP_WORDS[0]} abc" ];    then _picocli_mycmd; return $?; fi
+//        if [ "${COMP_LINE}" = "${COMP_WORDS[0]} abcdef" ]; then _picocli_mycmd; return $?; fi
+//        if [ "${COMP_LINE}" = "${COMP_WORDS[0]} generate-completion" ]; then _picocli_mycmd; return $?; fi
+//        if [ "${COMP_LINE}" = "${COMP_WORDS[0]} abcdef sub1" ]; then _picocli_mycmd_abcdef; return $?; fi
+//        if [ "${COMP_LINE}" = "${COMP_WORDS[0]} abcdef sub2" ]; then _picocli_mycmd_abcdef; return $?; fi
+//
 //                "  CMDS1=(%1$s gettingstarted)\n" +
 //                "  CMDS2=(%1$s tool)\n" +
 //                "  CMDS3=(%1$s tool sub1)\n" +
@@ -555,29 +605,39 @@ public class AutoComplete {
         StringBuilder buff = new StringBuilder(1024);
         buff.append(format(FUNCTION_HEADER, scriptName));
 
-        List<String> functionCallsToArrContains = new ArrayList<String>();
+        generatedEdgeCaseFunctionCalls(buff, hierarchy);
+        generateFunctionCallsToArrContains(buff, hierarchy);
 
-        generateFunctionCallsToArrContains(buff, functionCallsToArrContains, hierarchy);
-
-        buff.append("\n");
-        Collections.reverse(functionCallsToArrContains);
-        for (String func : functionCallsToArrContains) {
-            buff.append(func);
-        }
         buff.append(format(FUNCTION_FOOTER, scriptName));
         return buff.toString();
     }
 
-    private static void generateFunctionCallsToArrContains(StringBuilder buff,
-                                                           List<String> functionCalls,
-                                                           List<CommandDescriptor> hierarchy) {
+    // https://github.com/remkop/picocli/issues/1468
+    private static void generatedEdgeCaseFunctionCalls(StringBuilder buff, List<CommandDescriptor> hierarchy) {
+        buff.append("  # Edge case: if command line has no space after subcommand, then don't assume this subcommand is selected (remkop/picocli#1468).\n");
+        for (CommandDescriptor descriptor : hierarchy.subList(1, hierarchy.size())) { // skip top-level command
+            String withoutTopLevelCommand = concat(" ", descriptor.parentWithoutTopLevelCommand, descriptor.commandName);
+            buff.append(format("  if [ \"${COMP_LINE}\" = \"${COMP_WORDS[0]} %1$s\" ];    then %2$s; return $?; fi\n", withoutTopLevelCommand, descriptor.parentFunctionName));
+        }
+        buff.append("\n");
+    }
 
+    private static void generateFunctionCallsToArrContains(StringBuilder buff,
+                                                           List<CommandDescriptor> hierarchy) {
+        buff.append("  # Find the longest sequence of subcommands and call the bash function for that subcommand.\n");
+        List<String> functionCalls = new ArrayList<String>();
         for (CommandDescriptor descriptor : hierarchy.subList(1, hierarchy.size())) { // skip top-level command
             int count = functionCalls.size();
             String withoutTopLevelCommand = concat(" ", descriptor.parentWithoutTopLevelCommand, descriptor.commandName);
 
             functionCalls.add(format("  if CompWordsContainsArray \"${cmds%2$d[@]}\"; then %1$s; return $?; fi\n", descriptor.functionName, count));
             buff.append(      format("  local cmds%2$d=(%1$s)\n", withoutTopLevelCommand, count));
+        }
+
+        buff.append("\n");
+        Collections.reverse(functionCalls);
+        for (String func : functionCalls) {
+            buff.append(func);
         }
     }
     private static String concat(String infix, String... values) {
@@ -616,7 +676,8 @@ public class AutoComplete {
                 "  else\n" +
                 "    local positionals=\"\"\n" +
                 "%s" +
-                "    COMPREPLY=( $(compgen -W \"${commands} ${positionals}\" -- \"${curr_word}\") )\n" +
+                "    local IFS=$'\\n'\n" +
+                "    COMPREPLY=( $(compgen -W \"${commands// /$'\\n'}${IFS}${positionals}\" -- \"${curr_word}\") )\n" +
                 "  fi\n" +
                 "}\n";
 
@@ -635,7 +696,7 @@ public class AutoComplete {
             }
         }
         // If the command is a HelpCommand, append parent subcommands to the autocompletion list.
-        if (commandLine.getParent() != null && commandLine.getCommand() instanceof HelpCommand) {
+        if (commandLine.getParent() != null && ((Object) commandLine.getCommand()) instanceof HelpCommand) {
             subCommands = new LinkedHashSet<String>(subCommands);
             for (CommandLine subCommandLine : commandLine.getParent().getSubcommands().values()) {
                 if (commandLine == subCommandLine) { continue; }  // Skip the HelpCommand itself
@@ -690,16 +751,16 @@ public class AutoComplete {
 
     private static void generatePositionParamCompletionCandidates(StringBuilder buff, PositionalParamSpec f) {
         String paramName = bashify(f.paramLabel());
-        buff.append(format("  local %s_pos_param_args=\"%s\" # %d-%d values\n",
+        buff.append(format("  local %s_pos_param_args=(\"%s\") # %d-%d values\n",
                 paramName,
-                concat(" ", extract(f.completionCandidates())).trim(),
+                concat("\" \"", extract(f.completionCandidates())).trim(),
                 f.index().min(), f.index().max()));
     }
 
     private static void generateCompletionCandidates(StringBuilder buff, OptionSpec f) {
-        buff.append(format("  local %s_option_args=\"%s\" # %s values\n",
+        buff.append(format("  local %s_option_args=(\"%s\") # %s values\n",
                 bashify(f.paramLabel()),
-                concat(" ", extract(f.completionCandidates())).trim(),
+                concat("\" \"", extract(f.completionCandidates())).trim(),
                 f.longestName()));
     }
     private static List<String> extract(Iterable<String> generator) {
@@ -724,15 +785,15 @@ public class AutoComplete {
             int max = param.index().max();
             if (param.completionCandidates() != null) {
                 buff.append(format("%s    %s (( currIndex >= %d && currIndex <= %d )); then\n", indent, ifOrElif, min, max));
-                buff.append(format("%s      positionals=$( compgen -W \"$%s_pos_param_args\" -- \"%s\" )\n", indent, paramName, currWord));
+                buff.append(format("%s      positionals=$( compReplyArray \"${%s_pos_param_args[@]}\" )\n", indent, paramName));
             } else if (type.equals(File.class) || "java.nio.file.Path".equals(type.getName())) {
                 buff.append(format("%s    %s (( currIndex >= %d && currIndex <= %d )); then\n", indent, ifOrElif, min, max));
                 buff.append(format("%s      local IFS=$'\\n'\n", indent));
-                buff.append(format("%s      compopt -o filenames\n", indent));
+                buff.append(format("%s      type compopt &>/dev/null && compopt -o filenames\n", indent)); // #1464 workaround for old bash
                 buff.append(format("%s      positionals=$( compgen -f -- \"%s\" ) # files\n", indent, currWord));
             } else if (type.equals(InetAddress.class)) {
                 buff.append(format("%s    %s (( currIndex >= %d && currIndex <= %d )); then\n", indent, ifOrElif, min, max));
-                buff.append(format("%s      compopt -o filenames\n", indent));
+                buff.append(format("%s      type compopt &>/dev/null && compopt -o filenames\n", indent)); // #1464 workaround for old bash
                 buff.append(format("%s      positionals=$( compgen -A hostname -- \"%s\" )\n", indent, currWord));
             }
         }
@@ -750,7 +811,7 @@ public class AutoComplete {
         }
 
         return "\n"
-                + "  compopt +o default\n"
+                + "  type compopt &>/dev/null && compopt +o default\n" // #1464 workaround for old bash
                 + "\n"
                 + "  case ${prev_word} in\n"
                 + optionsCases
@@ -767,19 +828,20 @@ public class AutoComplete {
             }
             if (option.completionCandidates() != null) {
                 buff.append(format("%s    %s)\n", indent, concat("|", option.names()))); // "    -u|--timeUnit)\n"
-                buff.append(format("%s      COMPREPLY=( $( compgen -W \"${%s_option_args}\" -- \"%s\" ) )\n", indent, bashify(option.paramLabel()), currWord));
+                buff.append(format("%s      local IFS=$'\\n'\n", indent));
+                buff.append(format("%s      COMPREPLY=( $( compReplyArray \"${%s_option_args[@]}\" ) )\n", indent, bashify(option.paramLabel())));
                 buff.append(format("%s      return $?\n", indent));
                 buff.append(format("%s      ;;\n", indent));
             } else if (type.equals(File.class) || "java.nio.file.Path".equals(type.getName())) {
                 buff.append(format("%s    %s)\n", indent, concat("|", option.names()))); // "    -f|--file)\n"
                 buff.append(format("%s      local IFS=$'\\n'\n", indent));
-                buff.append(format("%s      compopt -o filenames\n", indent));
+                buff.append(format("%s      type compopt &>/dev/null && compopt -o filenames\n", indent)); // #1464 workaround for old bash
                 buff.append(format("%s      COMPREPLY=( $( compgen -f -- \"%s\" ) ) # files\n", indent, currWord));
                 buff.append(format("%s      return $?\n", indent));
                 buff.append(format("%s      ;;\n", indent));
             } else if (type.equals(InetAddress.class)) {
                 buff.append(format("%s    %s)\n", indent, concat("|", option.names()))); // "    -h|--host)\n"
-                buff.append(format("%s      compopt -o filenames\n", indent));
+                buff.append(format("%s      type compopt &>/dev/null && compopt -o filenames\n", indent)); // #1464 workaround for old bash
                 buff.append(format("%s      COMPREPLY=( $( compgen -A hostname -- \"%s\" ) )\n", indent, currWord));
                 buff.append(format("%s      return $?\n", indent));
                 buff.append(format("%s      ;;\n", indent));
