@@ -36,6 +36,7 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -96,21 +97,21 @@ public class CommandMethodTest {
     @SuppressWarnings("deprecation")
     @Test
     public void testAnnotateMethod_noArg() throws Exception {
-        setTraceLevel("OFF");
+        setTraceLevel(CommandLine.TraceLevel.OFF);
         Method m = CommandLine.getCommandMethods(MethodApp.class, "run0").get(0);
         CommandLine cmd1 = new CommandLine(m);
         assertEquals("run-0", cmd1.getCommandName());
-        assertEquals(Arrays.asList(), cmd1.getCommandSpec().args());
+        assertEquals(Collections.emptyList(), cmd1.getCommandSpec().args());
 
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
         cmd1.parseWithHandler(((IParseResultHandler) null), new PrintStream(baos), new String[]{"--y"});
-        assertEquals(Arrays.asList("--y"), cmd1.getUnmatchedArguments());
+        assertEquals(Collections.singletonList("--y"), cmd1.getUnmatchedArguments());
 
         // test execute
         Object ret = CommandLine.invoke(m.getName(), MethodApp.class, new PrintStream(new ByteArrayOutputStream()));
         assertNull("return value", ret);
 
-        setTraceLevel("WARN");
+        setTraceLevel(CommandLine.TraceLevel.WARN);
     }
     @Test
     public void testAnnotateMethod_unannotatedPositional() throws Exception {
@@ -450,7 +451,7 @@ public class CommandMethodTest {
     }
     @Test
     public void testAnnotateMethod_matchesAnnotatedClass() throws Exception {
-        setTraceLevel("OFF");
+        setTraceLevel(CommandLine.TraceLevel.OFF);
         CommandLine classCmd = new CommandLine(new CompactFields());
         Method m = CompactFieldsMethod.class.getDeclaredMethod("run", new Class<?>[] {boolean.class, boolean.class, File.class, File[].class});
         CommandLine methodCmd = new CommandLine(m);
@@ -461,7 +462,7 @@ public class CommandMethodTest {
             Model.ArgSpec methodArg = methodCmd.getCommandSpec().args().get(i);
             assertEquals("arg #" + i, classArg, methodArg);
         }
-        setTraceLevel("WARN");
+        setTraceLevel(CommandLine.TraceLevel.WARN);
     }
     /** replicate {@link CommandLineTest#testCompactFieldsAnyOrder()} but using
      * {@link CompactFieldsMethod#run(boolean, boolean, File, File[])}
