@@ -50,6 +50,7 @@ import java.util.concurrent.TimeUnit;
 import static java.lang.String.format;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.startsWith;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.*;
 import static picocli.CommandLine.*;
 import static picocli.CommandLine.Model.UsageMessageSpec.keyValuesMap;
@@ -832,7 +833,7 @@ public class ExecuteTest {
             @Option(names = "-x") int x;
             public void run() { }
         }
-        TestUtil.setTraceLevel("DEBUG");
+        TestUtil.setTraceLevel(CommandLine.TraceLevel.DEBUG);
         CommandLine cmd = new CommandLine(new App());
         cmd.execute("-x=abc");
 
@@ -918,7 +919,7 @@ public class ExecuteTest {
 
             public void run() { }
         }
-        TestUtil.setTraceLevel("DEBUG");
+        TestUtil.setTraceLevel(CommandLine.TraceLevel.DEBUG);
         CommandLine cmd = new CommandLine(new App());
         cmd.execute("abc");
 
@@ -1416,7 +1417,7 @@ public class ExecuteTest {
 
     @Test
     public void testKeyValuesMapIgnoresInvalidEntries() {
-        TestUtil.setTraceLevel("INFO");
+        TestUtil.setTraceLevel(CommandLine.TraceLevel.INFO);
         Map<String, String> map = keyValuesMap(" 0:Normal Execution",
                 "INVALID ENTRY",
                 "70:Internal error");
