@@ -1,5 +1,6 @@
 package picocli;
 
+import org.junit.After;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.contrib.java.lang.system.ProvideSystemProperty;
@@ -20,6 +21,7 @@ import java.util.concurrent.Callable;
 import static java.lang.String.format;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.not;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.*;
 
 public class InteractiveArgTest {
@@ -31,6 +33,66 @@ public class InteractiveArgTest {
     @Rule
     public final ProvideSystemProperty ansiOFF = new ProvideSystemProperty("picocli.ansi", "false");
 
+    static class Streams {
+        PrintStream out = System.out;
+        PrintStream err = System.err;
+        InputStream in = System.in;
+
+        void reset() {
+            System.setOut(out);
+            System.setOut(err);
+            System.setIn(in);
+        }
+    }
+
+    static class Capture {
+        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        ByteArrayOutputStream errBaos = new ByteArrayOutputStream();
+
+        Capture() {
+            this(null);
+        }
+
+        Capture(String input) {
+            System.setOut(new PrintStream(baos));
+            System.setErr(new PrintStream(errBaos));
+            if (input != null) {
+                System.setIn(new ByteArrayInputStream(input.getBytes()));
+            }
+        }
+
+        static Capture multiInputValue(String input) {
+            Capture result = new Capture(null);
+            System.setIn(inputStream(input));
+            return result;
+        }
+
+        private static ByteArrayInputStream inputStream(final String value) {
+            return new ByteArrayInputStream(value.getBytes()) {
+                int count;
+
+                @Override
+                public synchronized int read(byte[] b, int off, int len) {
+                    System.arraycopy(value.getBytes(), 0, b, off, value.length());
+                    return (count++ % 3) == 0 ? value.length() : -1;
+                }
+            };
+        }
+
+        String out() {
+            return baos.toString();
+        }
+
+        String err() {
+            return errBaos.toString();
+        }
+    }
+
+    @After
+    public void afterEachTest() {
+//        CommandLine.tracer().setLevel(CommandLine.TraceLevel.WARN);
+    }
+
     @Test
     public void testInteractiveOptionReadsFromStdIn() {
         class App {
@@ -38,16 +100,10 @@ public class InteractiveArgTest {
             @Option(names = "-z") int z;
         }
 
-        PrintStream out = System.out;
-        PrintStream err = System.err;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         System.setProperty("picocli.trace", "DEBUG");
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            ByteArrayOutputStream errBaos = new ByteArrayOutputStream();
-            System.setErr(new PrintStream(errBaos));
-            System.setIn(new ByteArrayInputStream("1234567890".getBytes()));
+            Capture capture = new Capture("1234567890");
 
             App app = new App();
             CommandLine cmd = new CommandLine(app);
@@ -55,14 +111,14 @@ public class InteractiveArgTest {
             ArgSpec specX = result.matchedArgs().get(0);
             assertThat(specX.toString(), containsString("App.x"));
 
-            assertEquals("Enter value for -x (Pwd): ", baos.toString());
+            assertEquals("Enter value for -x (Pwd): ", capture.out());
             assertEquals(1234567890, app.x);
             assertEquals(0, app.z);
 
-            String trace = errBaos.toString();
+            String trace = capture.err();
             assertThat(trace, containsString("User entered 10 characters"));
             assertThat(trace, containsString(
-                "Setting " + specX.toString() + " to *** (masked interactive value)"));
+                "Setting " + specX.toString() + " to *****(masked) (interactive value)"));
             assertThat(trace, not(containsString("1234567890")));
 
             cmd.parseArgs("-z", "678");
@@ -70,9 +126,7 @@ public class InteractiveArgTest {
             assertEquals(0, app.x);
             assertEquals(678, app.z);
         } finally {
-            System.setOut(out);
-            System.setOut(err);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -82,16 +136,10 @@ public class InteractiveArgTest {
             @Option(names = "-x", description = {"Pwd", "line2"}, interactive = true, echo = true) int x;
         }
 
-        PrintStream out = System.out;
-        PrintStream err = System.err;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         System.setProperty("picocli.trace", "DEBUG");
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            ByteArrayOutputStream errBaos = new ByteArrayOutputStream();
-            System.setErr(new PrintStream(errBaos));
-            System.setIn(new ByteArrayInputStream("1234567890".getBytes()));
+            Capture capture = new Capture("1234567890");
 
             App app = new App();
             CommandLine cmd = new CommandLine(app);
@@ -99,19 +147,38 @@ public class InteractiveArgTest {
             ArgSpec specX = result.matchedArgs().get(0);
             assertThat(specX.toString(), containsString("App.x"));
 
-            assertEquals("Enter value for -x (Pwd): ", baos.toString());
+            assertEquals("Enter value for -x (Pwd): ", capture.out());
             assertEquals(1234567890, app.x);
 
-            String trace = errBaos.toString();
+            String trace = capture.err();
             assertThat(trace, containsString("User entered 1234567890"));
             assertThat(trace, containsString(
                 "Setting " + specX.toString() + " to 1234567890"));
             assertThat(trace, not(containsString("10 characters")));
             assertThat(trace, not(containsString("***")));
         } finally {
-            System.setOut(out);
-            System.setErr(err);
-            System.setIn(in);
+            streams.reset();
+        }
+    }
+
+    @Test
+    public void testInteractiveOptionWithoutDescriptionStandardPrompt() {
+        class App {
+            @Option(names = "-x", interactive = true) int x;
+        }
+
+        Streams streams = new Streams();
+        try {
+            Capture capture = new Capture("123");
+
+            App app = new App();
+            CommandLine cmd = new CommandLine(app);
+            cmd.parseArgs("-x");
+
+            assertEquals("Enter value for -x: ", capture.out());
+            assertEquals(123, app.x);
+        } finally {
+            streams.reset();
         }
     }
 
@@ -121,22 +188,18 @@ public class InteractiveArgTest {
             @Option(names = "-x", description = {"Pwd", "line2"}, interactive = true, prompt = "[Customized]Enter your X: ") int x;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(new ByteArrayInputStream("123".getBytes()));
+            Capture capture = new Capture("123");
 
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x");
 
-            assertEquals("[Customized]Enter your X: ", baos.toString());
+            assertEquals("[Customized]Enter your X: ", capture.out());
             assertEquals(123, app.x);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -150,27 +213,21 @@ public class InteractiveArgTest {
             int z;
         }
 
-        PrintStream out = System.out;
-        PrintStream err = System.err;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         System.setProperty("picocli.trace", "DEBUG");
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            ByteArrayOutputStream errBaos = new ByteArrayOutputStream();
-            System.setErr(new PrintStream(errBaos));
-            System.setIn(inputStream("1234567890"));
+            Capture capture = Capture.multiInputValue("1234567890");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             ParseResult result = cmd.parseArgs("-x", "-x");
             ArgSpec specX = result.matchedArgs().get(0);
             assertThat(specX.toString(), containsString("App.x"));
 
-            assertEquals("Enter value for -x (Pwd): Enter value for -x (Pwd): ", baos.toString());
+            assertEquals("Enter value for -x (Pwd): Enter value for -x (Pwd): ", capture.out());
             assertEquals(Arrays.asList(1234567890, 1234567890), app.x);
             assertEquals(0, app.z);
 
-            String trace = errBaos.toString();
+            String trace = capture.err();
             assertThat(trace, containsString("User entered 10 characters"));
             assertThat(trace, containsString(
                 "Adding *** (masked interactive value) to " + specX.toString()
@@ -182,9 +239,7 @@ public class InteractiveArgTest {
             assertNull(app.x);
             assertEquals(678, app.z);
         } finally {
-            System.setOut(out);
-            System.setOut(err);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -198,27 +253,21 @@ public class InteractiveArgTest {
             int z;
         }
 
-        PrintStream out = System.out;
-        PrintStream err = System.err;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         System.setProperty("picocli.trace", "DEBUG");
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            ByteArrayOutputStream errBaos = new ByteArrayOutputStream();
-            System.setErr(new PrintStream(errBaos));
-            System.setIn(inputStream("1234567890"));
+            Capture capture = Capture.multiInputValue("1234567890");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             ParseResult result = cmd.parseArgs("-x", "-x");
             ArgSpec specX = result.matchedArgs().get(0);
             assertThat(specX.toString(), containsString("App.x"));
 
-            assertEquals("Enter value for -x (Pwd): Enter value for -x (Pwd): ", baos.toString());
+            assertEquals("Enter value for -x (Pwd): Enter value for -x (Pwd): ", capture.out());
             assertEquals(Arrays.asList(1234567890, 1234567890), app.x);
             assertEquals(0, app.z);
 
-            String trace = errBaos.toString();
+            String trace = capture.err();
             assertThat(trace, containsString("User entered 1234567890"));
             assertThat(trace, containsString(
                 "Adding 1234567890 (interactive value) to "
@@ -226,9 +275,7 @@ public class InteractiveArgTest {
             assertThat(trace, not(containsString("10 characters")));
             assertThat(trace, not(containsString("***")));
         } finally {
-            System.setOut(out);
-            System.setOut(err);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -242,35 +289,19 @@ public class InteractiveArgTest {
             int z;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(inputStream("123"));
+            Capture capture = Capture.multiInputValue("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x", "-x");
 
-            assertEquals("[Customized]Enter your x: [Customized]Enter your x: ", baos.toString());
+            assertEquals("[Customized]Enter your x: [Customized]Enter your x: ", capture.out());
             assertEquals(Arrays.asList(123, 123), app.x);
             assertEquals(0, app.z);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
-    }
-
-    ByteArrayInputStream inputStream(final String value) {
-        return new ByteArrayInputStream(value.getBytes()) {
-            int count;
-
-            @Override
-            public synchronized int read(byte[] b, int off, int len) {
-                System.arraycopy(value.getBytes(), 0, b, off, value.length());
-                return (count++ % 3) == 0 ? value.length() : -1;
-            }
-        };
     }
 
     @Test
@@ -283,17 +314,14 @@ public class InteractiveArgTest {
             int z;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(inputStream("123"));
+            Capture capture = Capture.multiInputValue("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x", "-x");
 
-            assertEquals("Enter value for -x (Pwd): Enter value for -x (Pwd): ", baos.toString());
+            assertEquals("Enter value for -x (Pwd): Enter value for -x (Pwd): ", capture.out());
             assertEquals(2, app.x.size());
             assertArrayEquals("123".toCharArray(), app.x.get(0));
             assertArrayEquals("123".toCharArray(), app.x.get(1));
@@ -304,8 +332,7 @@ public class InteractiveArgTest {
             assertNull(app.x);
             assertEquals(678, app.z);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -319,22 +346,18 @@ public class InteractiveArgTest {
             int z;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x");
 
-            assertEquals("Enter value for -x (Pwd): ", baos.toString());
+            assertEquals("Enter value for -x (Pwd): ", capture.out());
             assertArrayEquals("123".toCharArray(), app.x);
             assertEquals(0, app.z);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -348,11 +371,9 @@ public class InteractiveArgTest {
             String[] remainder;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            System.setOut(new PrintStream(new ByteArrayOutputStream()));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x", "456", "abc");
@@ -360,8 +381,7 @@ public class InteractiveArgTest {
             assertArrayEquals("456".toCharArray(), app.x);
             assertArrayEquals(new String[]{"abc"}, app.remainder);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -375,11 +395,9 @@ public class InteractiveArgTest {
             String[] remainder;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            System.setOut(new PrintStream(new ByteArrayOutputStream()));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x", "456", "-x", "789", "abc");
@@ -389,8 +407,7 @@ public class InteractiveArgTest {
             assertArrayEquals("789".toCharArray(), app.x.get(1));
             assertArrayEquals(new String[]{"abc"}, app.remainder);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -407,11 +424,9 @@ public class InteractiveArgTest {
             String[] remainder;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            System.setOut(new PrintStream(new ByteArrayOutputStream()));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x", "-z", "456", "abc");
@@ -420,8 +435,7 @@ public class InteractiveArgTest {
             assertEquals(456, app.z);
             assertArrayEquals(new String[]{"abc"}, app.remainder);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -438,11 +452,9 @@ public class InteractiveArgTest {
             String[] remainder;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            System.setOut(new PrintStream(new ByteArrayOutputStream()));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x", "-z", "456", "abc");
@@ -452,8 +464,7 @@ public class InteractiveArgTest {
             assertEquals(456, app.z);
             assertArrayEquals(new String[]{"abc"}, app.remainder);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -470,11 +481,9 @@ public class InteractiveArgTest {
             String[] remainder;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            System.setOut(new PrintStream(new ByteArrayOutputStream()));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x", "-y", "456", "abc");
@@ -483,8 +492,7 @@ public class InteractiveArgTest {
             assertEquals(0, app.z);
             assertArrayEquals(new String[]{"456", "abc"}, app.remainder);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -501,11 +509,9 @@ public class InteractiveArgTest {
             String[] remainder;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            System.setOut(new PrintStream(new ByteArrayOutputStream()));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x", "-y", "-x", "-w", "456", "abc");
@@ -516,8 +522,7 @@ public class InteractiveArgTest {
             assertEquals(0, app.z);
             assertArrayEquals(new String[]{"456", "abc"}, app.remainder);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -528,24 +533,20 @@ public class InteractiveArgTest {
             @Option(names = "-z") int z;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(new ByteArrayInputStream("123".getBytes()));
+            Capture capture = new Capture("123");
 
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("-x", "-z", "987");
 
             String expectedPrompt = format("Enter value for -x (Pwd%nline2): ");
-            assertEquals(expectedPrompt, baos.toString());
+            assertEquals(expectedPrompt, capture.out());
             assertEquals(123, app.x);
             assertEquals(987, app.z);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -556,16 +557,10 @@ public class InteractiveArgTest {
             @Parameters(index = "1") int z;
         }
 
-        PrintStream out = System.out;
-        PrintStream err = System.err;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         System.setProperty("picocli.trace", "DEBUG");
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            ByteArrayOutputStream errBaos = new ByteArrayOutputStream();
-            System.setErr(new PrintStream(errBaos));
-            System.setIn(new ByteArrayInputStream("1234567890".getBytes()));
+            Capture capture = new Capture("1234567890");
 
             App app = new App();
             CommandLine cmd = new CommandLine(app);
@@ -574,19 +569,17 @@ public class InteractiveArgTest {
             assertThat(specX.toString(), containsString("App.x"));
 
             String expectedPrompt = format("Enter value for position 0 (Pwd%nline2): ");
-            assertEquals(expectedPrompt, baos.toString());
+            assertEquals(expectedPrompt, capture.out());
             assertEquals(1234567890, app.x);
             assertEquals(987, app.z);
 
-            String trace = errBaos.toString();
+            String trace = capture.err();
             assertThat(trace, containsString("User entered 10 characters"));
             assertThat(trace, containsString(
-                "Setting " + specX.toString() + " to *** (masked interactive value)"));
+                "Setting " + specX.toString() + " to *****(masked) (interactive value)"));
             assertThat(trace, not(containsString("1234567890")));
         } finally {
-            System.setOut(out);
-            System.setOut(err);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -597,16 +590,10 @@ public class InteractiveArgTest {
             @Parameters(index = "1") int z;
         }
 
-        PrintStream out = System.out;
-        PrintStream err = System.err;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         System.setProperty("picocli.trace", "DEBUG");
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            ByteArrayOutputStream errBaos = new ByteArrayOutputStream();
-            System.setErr(new PrintStream(errBaos));
-            System.setIn(new ByteArrayInputStream("1234567890".getBytes()));
+            Capture capture = new Capture("1234567890");
 
             App app = new App();
             CommandLine cmd = new CommandLine(app);
@@ -615,20 +602,18 @@ public class InteractiveArgTest {
             assertThat(specX.toString(), containsString("App.x"));
 
             String expectedPrompt = format("Enter value for position 0 (Pwd%nline2): ");
-            assertEquals(expectedPrompt, baos.toString());
+            assertEquals(expectedPrompt, capture.out());
             assertEquals(1234567890, app.x);
             assertEquals(987, app.z);
 
-            String trace = errBaos.toString();
+            String trace = capture.err();
             assertThat(trace, containsString("User entered 1234567890"));
             assertThat(trace, containsString(
                 "Setting " + specX.toString() + " to 123"));
             assertThat(trace, not(containsString("10 characters")));
             assertThat(trace, not(containsString("***")));
         } finally {
-            System.setOut(out);
-            System.setOut(err);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -639,24 +624,20 @@ public class InteractiveArgTest {
             @Parameters(index = "1") int z;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(new ByteArrayInputStream("123".getBytes()));
+            Capture capture = new Capture("123");
 
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("987");
 
             String expectedPrompt = format("[Customized]Enter your value: ");
-            assertEquals(expectedPrompt, baos.toString());
+            assertEquals(expectedPrompt, capture.out());
             assertEquals(123, app.x);
             assertEquals(987, app.z);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -669,12 +650,9 @@ public class InteractiveArgTest {
             @Option(names = "-s") String str;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(new ByteArrayInputStream("123".getBytes()));
+            Capture capture = new Capture("123");
 
             App app = new App();
             CommandLine cmd = new CommandLine(app);
@@ -685,10 +663,9 @@ public class InteractiveArgTest {
             assertEquals("No value was read from console", 0, app.x);
             //String expectedPrompt = format("Enter value for position 0 (Pwd): ");
             String expectedPrompt = ""; // interactive arg was not prompted
-            assertEquals(expectedPrompt, baos.toString());
+            assertEquals(expectedPrompt, capture.out());
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -700,25 +677,21 @@ public class InteractiveArgTest {
             @Parameters(index = "2") int z;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(new ByteArrayInputStream("123".getBytes()));
+            Capture capture = new Capture("123");
 
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("333", "987");
 
             String expectedPrompt = format("Enter value for position 1 (Pwd%nline2): ");
-            assertEquals(expectedPrompt, baos.toString());
+            assertEquals(expectedPrompt, capture.out());
             assertEquals(333, app.a);
             assertEquals(123, app.x);
             assertEquals(987, app.z);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -759,24 +732,20 @@ public class InteractiveArgTest {
             }
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(inputStream("password123"));
+            Capture capture = new Capture("password123");
 
             Login login = new Login();
             new CommandLine(login).execute("-u", "user123", "-p");
 
             String expectedPrompt = format("Enter value for --password (Password or passphrase): " +
                     "Hi user123, your password is hashed to 75K3eLr+dx6JJFuJ7LwIpEpOFmwGZZkRiB84PURz6U8=.%n");
-            assertEquals(expectedPrompt, baos.toString());
+            assertEquals(expectedPrompt, capture.out());
             assertEquals("user123", login.user);
             assertArrayEquals("           ".toCharArray(), login.password);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -788,24 +757,20 @@ public class InteractiveArgTest {
             @Parameters(index = "2") int z;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(inputStream("123"));
+            Capture capture = Capture.multiInputValue("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("999");
 
-            assertEquals("Enter value for position 0 (Pwd): Enter value for position 1 (Pwd): ", baos.toString());
+            assertEquals("Enter value for position 0 (Pwd): Enter value for position 1 (Pwd): ", capture.out());
             assertEquals(2, app.x.size());
             assertArrayEquals("123".toCharArray(), app.x.get(0));
             assertArrayEquals("123".toCharArray(), app.x.get(1));
             assertEquals(999, app.z);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -817,22 +782,18 @@ public class InteractiveArgTest {
             @Parameters(index = "1") int z;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-            System.setOut(new PrintStream(baos));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("9");
 
-            assertEquals("Enter value for position 0 (Pwd): ", baos.toString());
+            assertEquals("Enter value for position 0 (Pwd): ", capture.out());
             assertArrayEquals("123".toCharArray(), app.x);
             assertEquals(9, app.z);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -846,11 +807,9 @@ public class InteractiveArgTest {
             String[] remainder;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            System.setOut(new PrintStream(new ByteArrayOutputStream()));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("456", "abc");
@@ -858,8 +817,7 @@ public class InteractiveArgTest {
             assertArrayEquals("456".toCharArray(), app.x);
             assertArrayEquals(new String[]{"abc"}, app.remainder);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -873,11 +831,9 @@ public class InteractiveArgTest {
             String[] remainder;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            System.setOut(new PrintStream(new ByteArrayOutputStream()));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             cmd.parseArgs("456", "789", "abc");
@@ -887,8 +843,7 @@ public class InteractiveArgTest {
             assertArrayEquals("789".toCharArray(), app.x.get(1));
             assertArrayEquals(new String[]{"abc"}, app.remainder);
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -905,11 +860,9 @@ public class InteractiveArgTest {
             String[] remainder;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            System.setOut(new PrintStream(new ByteArrayOutputStream()));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             try {
@@ -919,8 +872,7 @@ public class InteractiveArgTest {
                 assertEquals("Unknown option: '-y'", ex.getMessage());
             }
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
         }
     }
 
@@ -937,11 +889,9 @@ public class InteractiveArgTest {
             String[] remainder;
         }
 
-        PrintStream out = System.out;
-        InputStream in = System.in;
+        Streams streams = new Streams();
         try {
-            System.setOut(new PrintStream(new ByteArrayOutputStream()));
-            System.setIn(inputStream("123"));
+            Capture capture = new Capture("123");
             App app = new App();
             CommandLine cmd = new CommandLine(app);
             try {
@@ -951,8 +901,189 @@ public class InteractiveArgTest {
                 assertEquals("Unknown options: '-y', '-w'", ex.getMessage());
             }
         } finally {
-            System.setOut(out);
-            System.setIn(in);
+            streams.reset();
+        }
+    }
+
+    @Test
+    public void testIssue2087InteractiveEchoOffOptionWithDefault() {
+        class Issue2087a {
+            @Option(names = "-a", interactive = true, echo = false, defaultValue = "DEFAULT") String a;
+        }
+
+        System.setProperty("picocli.trace", "DEBUG");
+        Streams streams = new Streams();
+        try {
+            Capture capture = new Capture();
+
+            Issue2087a app = new Issue2087a();
+            new CommandLine(app).parseArgs();
+            assertEquals("DEFAULT", app.a);
+
+            assertEquals("", capture.out());
+            assertThat(capture.err(), containsString("Applying defaultValue (*****(masked)) to field String picocli.InteractiveArgTest$1Issue2087a.a"));
+            assertThat(capture.err(), containsString("Setting field String picocli.InteractiveArgTest$1Issue2087a.a to '*****(masked)' (was 'null') for"));
+        } finally {
+            streams.reset();
+        }
+    }
+
+    @Test
+    public void testIssue2087InteractiveEchoOnOptionWithDefault() {
+        class Issue2087b {
+            @Option(names = "-a", interactive = true, echo = true, defaultValue = "DEFAULT") String a;
+        }
+
+        System.setProperty("picocli.trace", "DEBUG");
+        Streams streams = new Streams();
+        try {
+            Capture capture = new Capture();
+
+            Issue2087b app = new Issue2087b();
+            new CommandLine(app).parseArgs();
+            assertEquals("DEFAULT", app.a);
+
+            assertEquals("", capture.out());
+            assertThat(capture.err(), containsString("Applying defaultValue (DEFAULT) to field String picocli.InteractiveArgTest$1Issue2087b.a"));
+            assertThat(capture.err(), containsString("Setting field String picocli.InteractiveArgTest$1Issue2087b.a to 'DEFAULT' (was 'null') for"));
+        } finally {
+            streams.reset();
+        }
+    }
+
+    static class Issue2087DefaultProvider1 implements CommandLine.IDefaultValueProvider {
+        public String defaultValue(ArgSpec argSpec) throws Exception {
+            return "DEFAULT1";
+        }
+    }
+
+    @Test
+    public void testIssue2087InteractiveOptionEchoOffWithDefaultProvider() {
+        @CommandLine.Command(defaultValueProvider = Issue2087DefaultProvider1.class)
+        class Issue2087c {
+            @Option(names = "-a", interactive = true, echo = false) String a;
+        }
+        System.setProperty("picocli.trace", "DEBUG");
+        Streams streams = new Streams();
+        try {
+            Capture capture = new Capture();
+            Issue2087c app = new Issue2087c();
+            new CommandLine(app).parseArgs();
+            assertEquals("DEFAULT1", app.a);
+
+            assertEquals("", capture.out());
+            assertThat(capture.err(), containsString("Applying defaultValue (*****(masked)) from picocli.InteractiveArgTest$Issue2087DefaultProvider1"));
+            assertThat(capture.err(), containsString("Setting field String picocli.InteractiveArgTest$1Issue2087c.a to '*****(masked)' (was 'null') for"));
+        } finally {
+            streams.reset();
+        }
+    }
+
+    @Test
+    public void testIssue2087InteractiveOptionEchoOnWithDefaultProvider() {
+        @CommandLine.Command(defaultValueProvider = Issue2087DefaultProvider1.class)
+        class Issue2087d {
+            @Option(names = "-a", interactive = true, echo = true) String a;
+        }
+        System.setProperty("picocli.trace", "DEBUG");
+        Streams streams = new Streams();
+        try {
+            Capture capture = new Capture();
+            Issue2087d app = new Issue2087d();
+            new CommandLine(app).parseArgs();
+            assertEquals("DEFAULT1", app.a);
+
+            assertEquals("", capture.out());
+            assertThat(capture.err(), containsString("Applying defaultValue (DEFAULT1) from picocli.InteractiveArgTest$Issue2087DefaultProvider1"));
+            assertThat(capture.err(), containsString("Setting field String picocli.InteractiveArgTest$1Issue2087d.a to 'DEFAULT1' (was 'null') for"));
+        } finally {
+            streams.reset();
+        }
+    }
+
+    @Test
+    public void testIssue2087InteractiveEchoOnPositionalWithDefault() {
+        class Issue2087e {
+            @Parameters(interactive = true, echo = true, defaultValue = "DEFAULT") String a;
+        }
+        System.setProperty("picocli.trace", "DEBUG");
+        Streams streams = new Streams();
+        try {
+            Capture capture = new Capture();
+            Issue2087e app = new Issue2087e();
+            new CommandLine(app).parseArgs();
+            assertEquals("DEFAULT", app.a);
+
+            assertEquals("", capture.out());
+            assertThat(capture.err(), containsString("Applying defaultValue (DEFAULT) "));
+            assertThat(capture.err(), containsString("Setting field String picocli.InteractiveArgTest$1Issue2087e.a to 'DEFAULT' (was 'null') for"));
+        } finally {
+            streams.reset();
+        }
+    }
+
+    @Test
+    public void testIssue2087InteractiveEchoOffPositionalWithDefault() {
+        class Issue2087g {
+            @Parameters(interactive = true, echo = false, defaultValue = "DEFAULT") String a;
+        }
+        System.setProperty("picocli.trace", "DEBUG");
+        Streams streams = new Streams();
+        try {
+            Capture capture = new Capture();
+            Issue2087g app = new Issue2087g();
+            new CommandLine(app).parseArgs();
+            assertEquals("DEFAULT", app.a);
+
+            assertEquals("", capture.out());
+            assertThat(capture.err(), containsString("Applying defaultValue (*****(masked)) "));
+            assertThat(capture.err(), containsString("Setting field String picocli.InteractiveArgTest$1Issue2087g.a to '*****(masked)' (was 'null') for"));
+        } finally {
+            streams.reset();
+        }
+    }
+
+    @Test
+    public void testIssue2087InteractiveEchoOnPositionalWithDefaultProvider() {
+        @CommandLine.Command(defaultValueProvider = Issue2087DefaultProvider1.class)
+        class Issue2087f {
+            @Parameters(interactive = true, echo = true) String a;
+        }
+        System.setProperty("picocli.trace", "DEBUG");
+        Streams streams = new Streams();
+        try {
+            Capture capture = new Capture();
+            Issue2087f app = new Issue2087f();
+            new CommandLine(app).parseArgs();
+            assertEquals("DEFAULT1", app.a);
+
+            assertEquals("", capture.out());
+            assertThat(capture.err(), containsString("Applying defaultValue (DEFAULT1) from picocli.InteractiveArgTest$Issue2087DefaultProvider1"));
+            assertThat(capture.err(), containsString("Setting field String picocli.InteractiveArgTest$1Issue2087f.a to 'DEFAULT1' (was 'null') for"));
+        } finally {
+            streams.reset();
+        }
+    }
+
+    @Test
+    public void testIssue2087InteractiveEchoOffPositionalWithDefaultProvider() {
+        @CommandLine.Command(defaultValueProvider = Issue2087DefaultProvider1.class)
+        class Issue2087h {
+            @Parameters(interactive = true, echo = false) String a;
+        }
+        System.setProperty("picocli.trace", "DEBUG");
+        Streams streams = new Streams();
+        try {
+            Capture capture = new Capture();
+            Issue2087h app = new Issue2087h();
+            new CommandLine(app).parseArgs();
+            assertEquals("DEFAULT1", app.a);
+
+            assertEquals("", capture.out());
+            assertThat(capture.err(), containsString("Applying defaultValue (*****(masked)) from picocli.InteractiveArgTest$Issue2087DefaultProvider1"));
+            assertThat(capture.err(), containsString("Setting field String picocli.InteractiveArgTest$1Issue2087h.a to '*****(masked)' (was 'null') for"));
+        } finally {
+            streams.reset();
         }
     }
 
