@@ -4,6 +4,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.contrib.java.lang.system.ProvideSystemProperty;
 import org.junit.contrib.java.lang.system.RestoreSystemProperties;
+import org.junit.contrib.java.lang.system.SystemErrRule;
 import org.junit.rules.TestRule;
 
 import java.io.ByteArrayOutputStream;
@@ -22,6 +23,9 @@ import static picocli.TestUtil.stripHashcodes;
 
 public class TracerTest {
     @Rule
+    public final SystemErrRule systemErrRule = new SystemErrRule().enableLog().muteForSuccessfulTests();
+
+    @Rule
     public final TestRule restoreSystemProperties = new RestoreSystemProperties();
 
     @Rule
@@ -36,19 +40,8 @@ public class TracerTest {
     @Test
     public void testDebugOutputForDoubleDashSeparatesPositionalParameters() throws Exception {
         clearBuiltInTracingCache();
-        PrintStream originalErr = System.err;
-        ByteArrayOutputStream baos = new ByteArrayOutputStream(2500);
-        System.setErr(new PrintStream(baos));
-        final String PROPERTY = "picocli.trace";
-        String old = System.getProperty(PROPERTY);
-        System.setProperty(PROPERTY, "DEBUG");
+        TestUtil.setTraceLevel(CommandLine.TraceLevel.DEBUG);
         CommandLine.populateCommand(new CommandLineTest.CompactFields(), "-oout -- -r -v p1 p2".split(" "));
-        System.setErr(originalErr);
-        if (old == null) {
-            System.clearProperty(PROPERTY);
-        } else {
-            System.setProperty(PROPERTY, old);
-        }
         String prefix8 = format("" +
                 "[picocli DEBUG] Could not register converter for java.time.Duration: java.lang.ClassNotFoundException: java.time.Duration%n" +
                 "[picocli DEBUG] Could not register converter for java.time.Instant: java.lang.ClassNotFoundException: java.time.Instant%n" +
@@ -70,7 +63,7 @@ public class TracerTest {
                         "[picocli DEBUG] Creating CommandSpec for picocli.CommandLineTest$CompactFields@20f5239f with factory picocli.CommandLine$DefaultFactory%n" +
                         "[picocli INFO] Picocli version: %3$s%n" +
                         "[picocli INFO] Parsing 6 command line args [-oout, --, -r, -v, p1, p2]%n" +
-                        "[picocli DEBUG] Parser configuration: optionsCaseInsensitive=false, subcommandsCaseInsensitive=false, abbreviatedOptionsAllowed=false, abbreviatedSubcommandsAllowed=false, aritySatisfiedByAttachedOptionParam=false, atFileCommentChar=#, caseInsensitiveEnumValuesAllowed=false, collectErrors=false, endOfOptionsDelimiter=--, expandAtFiles=true, limitSplit=false, overwrittenOptionsAllowed=false, posixClusteredShortOptionsAllowed=true, separator=null, splitQuotedStrings=false, stopAtPositional=false, stopAtUnmatched=false, toggleBooleanFlags=false, trimQuotes=false, unmatchedArgumentsAllowed=false, unmatchedOptionsAllowedAsOptionParameters=true, unmatchedOptionsArePositionalParams=false, useSimplifiedAtFiles=false%n" +
+                        "[picocli DEBUG] Parser configuration: optionsCaseInsensitive=false, subcommandsCaseInsensitive=false, abbreviatedOptionsAllowed=false, abbreviatedSubcommandsAllowed=false, allowOptionsAsOptionParameters=false, allowSubcommandsAsOptionParameters=false, aritySatisfiedByAttachedOptionParam=false, atFileCommentChar=#, caseInsensitiveEnumValuesAllowed=false, collectErrors=false, endOfOptionsDelimiter=--, expandAtFiles=true, limitSplit=false, overwrittenOptionsAllowed=false, posixClusteredShortOptionsAllowed=true, separator=null, splitQuotedStrings=false, stopAtPositional=false, stopAtUnmatched=false, toggleBooleanFlags=false, trimQuotes=false, unmatchedArgumentsAllowed=false, unmatchedOptionsAllowedAsOptionParameters=true, unmatchedOptionsArePositionalParams=false, useSimplifiedAtFiles=false%n" +
                         "[picocli DEBUG] (ANSI is disabled by default: ...)%n" +
                         "[picocli DEBUG] Initializing command 'null' (user object: picocli.CommandLineTest$CompactFields@20f5239f): 3 options, 1 positional parameters, 0 required, 0 groups, 0 subcommands.%n" +
                         "[picocli DEBUG] Set initial value for field boolean picocli.CommandLineTest$CompactFields.verbose of type boolean to false.%n" +
@@ -114,7 +107,7 @@ public class TracerTest {
                 CommandLineTest.class.getName(),
                 new File("/home/rpopma/picocli"),
                 CommandLine.versionString());
-        String actual = new String(baos.toByteArray(), "UTF8");
+        String actual = systemErrRule.getLog();
         //System.out.println(actual);
         if (System.getProperty("java.version").compareTo("1.7.0") < 0) {
             expected = prefix7 + expected;
@@ -132,13 +125,13 @@ public class TracerTest {
 
         try {
             System.clearProperty(PROPERTY);
-            assertTrue("WARN enabled by default", new CommandLine.Tracer().isWarn());
+            assertTrue("WARN enabled by default", CommandLine.tracer().isWarn());
 
             System.setProperty(PROPERTY, "OFF");
-            assertFalse("WARN can be disabled by setting to OFF", new CommandLine.Tracer().isWarn());
+            assertFalse("WARN can be disabled by setting to OFF", CommandLine.tracer().isWarn());
 
             System.setProperty(PROPERTY, "WARN");
-            assertTrue("WARN can be explicitly enabled", new CommandLine.Tracer().isWarn());
+            assertTrue("WARN can be explicitly enabled", CommandLine.tracer().isWarn());
 
         } finally {
             if (old == null) {
@@ -156,28 +149,28 @@ public class TracerTest {
 
         try {
             System.setProperty(PROPERTY, "off");
-            assertEquals("OFF", String.valueOf(new CommandLine.Tracer().level));
-            assertFalse("!debug", new CommandLine.Tracer().isDebug());
-            assertFalse("!info", new CommandLine.Tracer().isInfo());
-            assertFalse("!warn", new CommandLine.Tracer().isWarn());
+            assertEquals("OFF", String.valueOf(CommandLine.tracer().getLevel()));
+            assertFalse("!debug", CommandLine.tracer().isDebug());
+            assertFalse("!info", CommandLine.tracer().isInfo());
+            assertFalse("!warn", CommandLine.tracer().isWarn());
 
             System.setProperty(PROPERTY, "debug");
-            assertEquals("DEBUG", String.valueOf(new CommandLine.Tracer().level));
-            assertTrue("debug", new CommandLine.Tracer().isDebug());
-            assertTrue("info", new CommandLine.Tracer().isInfo());
-            assertTrue("warn", new CommandLine.Tracer().isWarn());
+            assertEquals("DEBUG", String.valueOf(CommandLine.tracer().getLevel()));
+            assertTrue("debug", CommandLine.tracer().isDebug());
+            assertTrue("info", CommandLine.tracer().isInfo());
+            assertTrue("warn", CommandLine.tracer().isWarn());
 
             System.setProperty(PROPERTY, "info");
-            assertEquals("INFO", String.valueOf(new CommandLine.Tracer().level));
-            assertFalse("!debug", new CommandLine.Tracer().isDebug());
-            assertTrue("info", new CommandLine.Tracer().isInfo());
-            assertTrue("warn", new CommandLine.Tracer().isWarn());
+            assertEquals("INFO", String.valueOf(CommandLine.tracer().getLevel()));
+            assertFalse("!debug", CommandLine.tracer().isDebug());
+            assertTrue("info", CommandLine.tracer().isInfo());
+            assertTrue("warn", CommandLine.tracer().isWarn());
 
             System.setProperty(PROPERTY, "warn");
-            assertEquals("WARN", String.valueOf(new CommandLine.Tracer().level));
-            assertFalse("!debug", new CommandLine.Tracer().isDebug());
-            assertFalse("!info", new CommandLine.Tracer().isInfo());
-            assertTrue("warn", new CommandLine.Tracer().isWarn());
+            assertEquals("WARN", String.valueOf(CommandLine.tracer().getLevel()));
+            assertFalse("!debug", CommandLine.tracer().isDebug());
+            assertFalse("!info", CommandLine.tracer().isInfo());
+            assertTrue("warn", CommandLine.tracer().isWarn());
 
         } finally {
             if (old == null) {
@@ -189,17 +182,38 @@ public class TracerTest {
     }
 
     @Test
+    public void testTraceLevelToInfoViaSetterWithSubCommands() throws Exception {
+        CommandLine.TraceLevel old = CommandLine.tracer().getLevel();
+        CommandLine.tracer().setLevel(CommandLine.TraceLevel.INFO);
+        CommandLine commandLine = Demo.mainCommand();
+        commandLine.setEndOfOptionsDelimiter("$$$");
+        commandLine.parseArgs("--git-dir=/home/rpopma/picocli", "commit", "-m", "\"Fixed typos\"", "$$$", "src1.java", "src2.java", "src3.java");
+        CommandLine.tracer().setLevel(old);
+        CommandLine.tracer().modified = false;
+        String expected = format("" +
+                "[picocli INFO] Picocli version: %s%n" +
+                "[picocli INFO] Parsing 8 command line args [--git-dir=/home/rpopma/picocli, commit, -m, \"Fixed typos\", $$$, src1.java, src2.java, src3.java]%n" +
+                "[picocli INFO] Setting field java.io.File picocli.Demo$Git.gitDir to '%s' (was 'null') for option --git-dir%n" +
+                "[picocli INFO] Adding [\"Fixed typos\"] to field java.util.List<String> picocli.Demo$GitCommit.message for option -m%n" +
+                "[picocli INFO] Found end-of-options delimiter '$$$'. Treating remainder as positional parameters.%n" +
+                "[picocli INFO] Adding [src1.java] to field java.util.List<java.io.File> picocli.Demo$GitCommit.files for args[0..*] at position 0%n" +
+                "[picocli INFO] Adding [src2.java] to field java.util.List<java.io.File> picocli.Demo$GitCommit.files for args[0..*] at position 1%n" +
+                "[picocli INFO] Adding [src3.java] to field java.util.List<java.io.File> picocli.Demo$GitCommit.files for args[0..*] at position 2%n",
+            CommandLine.versionString(),
+            new File("/home/rpopma/picocli"));
+        String actual = systemErrRule.getLog();
+        //System.out.println(actual);
+        assertEquals(stripAnsiTrace(expected), stripAnsiTrace(actual));
+    }
+
+    @Test
     public void testTracingInfoWithSubCommands() throws Exception {
-        PrintStream originalErr = System.err;
-        ByteArrayOutputStream baos = new ByteArrayOutputStream(2500);
-        System.setErr(new PrintStream(baos));
         final String PROPERTY = "picocli.trace";
         String old = System.getProperty(PROPERTY);
         System.setProperty(PROPERTY, "");
         CommandLine commandLine = Demo.mainCommand();
         commandLine.setEndOfOptionsDelimiter("$$$");
         commandLine.parseArgs("--git-dir=/home/rpopma/picocli", "commit", "-m", "\"Fixed typos\"", "$$$", "src1.java", "src2.java", "src3.java");
-        System.setErr(originalErr);
         if (old == null) {
             System.clearProperty(PROPERTY);
         } else {
@@ -216,22 +230,18 @@ public class TracerTest {
                         "[picocli INFO] Adding [src3.java] to field java.util.List<java.io.File> picocli.Demo$GitCommit.files for args[0..*] at position 2%n",
                 CommandLine.versionString(),
                 new File("/home/rpopma/picocli"));
-        String actual = new String(baos.toByteArray(), "UTF8");
+        String actual = systemErrRule.getLog();
         //System.out.println(actual);
         assertEquals(stripAnsiTrace(expected), stripAnsiTrace(actual));
     }
     @Test
     public void testTracingDebugWithSubCommands() throws Exception {
         clearBuiltInTracingCache();
-        PrintStream originalErr = System.err;
-        ByteArrayOutputStream baos = new ByteArrayOutputStream(2500);
-        System.setErr(new PrintStream(baos));
         final String PROPERTY = "picocli.trace";
         String old = System.getProperty(PROPERTY);
         System.setProperty(PROPERTY, "DEBUG");
         CommandLine commandLine = Demo.mainCommand();
-        commandLine.parseArgs("--git-dir=/home/rpopma/picocli", "commit", "-m", "\"Fixed typos\"", "--", "src1.java", "src2.java", "src3.java");
-        System.setErr(originalErr);
+        commandLine.execute("--git-dir=/home/rpopma/picocli", "commit", "-m", "\"Fixed typos\"", "--", "src1.java", "src2.java", "src3.java");
         if (old == null) {
             System.clearProperty(PROPERTY);
         } else {
@@ -283,7 +293,7 @@ public class TracerTest {
                         "[picocli DEBUG] Adding subcommand 'tag' to 'git'%n" +
                         "[picocli INFO] Picocli version: %3$s%n" +
                         "[picocli INFO] Parsing 8 command line args [--git-dir=/home/rpopma/picocli, commit, -m, \"Fixed typos\", --, src1.java, src2.java, src3.java]%n" +
-                        "[picocli DEBUG] Parser configuration: optionsCaseInsensitive=false, subcommandsCaseInsensitive=false, abbreviatedOptionsAllowed=false, abbreviatedSubcommandsAllowed=false, aritySatisfiedByAttachedOptionParam=false, atFileCommentChar=#, caseInsensitiveEnumValuesAllowed=false, collectErrors=false, endOfOptionsDelimiter=--, expandAtFiles=true, limitSplit=false, overwrittenOptionsAllowed=false, posixClusteredShortOptionsAllowed=true, separator=null, splitQuotedStrings=false, stopAtPositional=false, stopAtUnmatched=false, toggleBooleanFlags=false, trimQuotes=false, unmatchedArgumentsAllowed=false, unmatchedOptionsAllowedAsOptionParameters=true, unmatchedOptionsArePositionalParams=false, useSimplifiedAtFiles=false%n" +
+                        "[picocli DEBUG] Parser configuration: optionsCaseInsensitive=false, subcommandsCaseInsensitive=false, abbreviatedOptionsAllowed=false, abbreviatedSubcommandsAllowed=false, allowOptionsAsOptionParameters=false, allowSubcommandsAsOptionParameters=false, aritySatisfiedByAttachedOptionParam=false, atFileCommentChar=#, caseInsensitiveEnumValuesAllowed=false, collectErrors=false, endOfOptionsDelimiter=--, expandAtFiles=true, limitSplit=false, overwrittenOptionsAllowed=false, posixClusteredShortOptionsAllowed=true, separator=null, splitQuotedStrings=false, stopAtPositional=false, stopAtUnmatched=false, toggleBooleanFlags=false, trimQuotes=false, unmatchedArgumentsAllowed=false, unmatchedOptionsAllowedAsOptionParameters=true, unmatchedOptionsArePositionalParams=false, useSimplifiedAtFiles=false%n" +
                         "[picocli DEBUG] (ANSI is disabled by default: ...)%n" +
                         "[picocli DEBUG] Initializing command 'git' (user object: picocli.Demo$Git@75d4a5c2): 3 options, 0 positional parameters, 0 required, 0 groups, 12 subcommands.%n" +
                         "[picocli DEBUG] Set initial value for field java.io.File picocli.Demo$Git.gitDir of type class java.io.File to null.%n" +
@@ -341,11 +351,19 @@ public class TracerTest {
                         "[picocli DEBUG] defaultValue not defined for field java.io.File picocli.Demo$GitCommit.file%n" +
                         "[picocli DEBUG] Applying default values for command 'git'%n" +
                         "[picocli DEBUG] defaultValue not defined for field boolean picocli.CommandLine$AutoHelpMixin.helpRequested%n" +
-                        "[picocli DEBUG] defaultValue not defined for field boolean picocli.CommandLine$AutoHelpMixin.versionRequested%n",
+                        "[picocli DEBUG] defaultValue not defined for field boolean picocli.CommandLine$AutoHelpMixin.versionRequested%n" +
+                "[picocli DEBUG] Help was not requested. Continuing to process ParseResult...%n" +
+                "[picocli DEBUG] RunLast: handling ParseResult...%n" +
+                "[picocli DEBUG] RunLast: executing user object for 'git git-commit'...%n" +
+                "[picocli DEBUG] Invoking Runnable::run%n" +
+                "[picocli DEBUG] RunLast: ParseResult has 0 exit code generators%n" +
+                "[picocli DEBUG] resolveExitCode: exit code generators resulted in exit code=0%n" +
+                "[picocli DEBUG] resolveExitCode: execution results resulted in exit code=0%n" +
+                "[picocli DEBUG] resolveExitCode: returning exit code=0%n",
                 Demo.class.getName(),
                 new File("/home/rpopma/picocli"),
                 CommandLine.versionString());
-        String actual = new String(baos.toByteArray(), "UTF8");
+        String actual = systemErrRule.getLog();
         String actualSafe = stripHashcodes(actual);
         //System.out.println(actual);
         if (System.getProperty("java.version").compareTo("1.7.0") < 0) {
@@ -389,11 +407,8 @@ public class TracerTest {
 
     @Test
     public void testTraceWarningIfOptionOverwrittenWhenOverwrittenOptionsAllowed() throws Exception {
-        PrintStream originalErr = System.err;
-        ByteArrayOutputStream baos = new ByteArrayOutputStream(2500);
-        System.setErr(new PrintStream(baos));
 
-        setTraceLevel("INFO");
+        setTraceLevel(CommandLine.TraceLevel.INFO);
         class App {
             @CommandLine.Option(names = "-f") String field = null;
             @CommandLine.Option(names = "-p") int primitive = 43;
@@ -402,7 +417,6 @@ public class TracerTest {
         cmd.parseArgs("-f", "111", "-f", "222", "-f", "333");
         App ff = cmd.getCommand();
         assertEquals("333", ff.field);
-        System.setErr(originalErr);
 
         String expected = format("" +
                         "[picocli INFO] Picocli version: %s%n" +
@@ -412,25 +426,20 @@ public class TracerTest {
                         "[picocli INFO] Overwriting field String %2$s.field value '222' with '333' for option -f%n",
                 CommandLine.versionString(),
                 App.class.getName());
-        String actual = new String(baos.toByteArray(), "UTF8");
+        String actual = systemErrRule.getLog();
         assertEquals(stripAnsiTrace(expected), stripAnsiTrace(actual));
-        setTraceLevel("WARN");
+        setTraceLevel(CommandLine.TraceLevel.WARN);
     }
     @SuppressWarnings("deprecation")
     @Test
     public void testTraceWarningIfUnmatchedArgsWhenUnmatchedArgumentsAllowed() throws Exception {
-        PrintStream originalErr = System.err;
-        ByteArrayOutputStream baos = new ByteArrayOutputStream(2500);
-        System.setErr(new PrintStream(baos));
-
-        setTraceLevel("INFO");
+        setTraceLevel(CommandLine.TraceLevel.INFO);
         class App {
             @CommandLine.Parameters(index = "0", arity = "2", split = "\\|", type = {Integer.class, String.class})
             Map<Integer,String> message;
         }
         CommandLine cmd = new CommandLine(new App()).setUnmatchedArgumentsAllowed(true).parse("1=a", "2=b", "3=c", "4=d").get(0);
         assertEquals(Arrays.asList("3=c", "4=d"), cmd.getUnmatchedArguments());
-        System.setErr(originalErr);
 
         String expected = format("" +
                         "[picocli INFO] Picocli version: %s%n" +
@@ -441,9 +450,9 @@ public class TracerTest {
                 CommandLine.versionString(),
                 App.class.getName(),
                 App.class.getName());
-        String actual = new String(baos.toByteArray(), "UTF8");
+        String actual = systemErrRule.getLog();
         //System.out.println(actual);
         assertEquals(stripAnsiTrace(expected), stripAnsiTrace(actual));
-        setTraceLevel("WARN");
+        setTraceLevel(CommandLine.TraceLevel.WARN);
     }
 }
